Make readiness endpoint & probe more sophisticated
The probe in the makefile will now fail on 5XX errors (as it should).

The readiness endpoint will test for the tables it requires in the
database as well as database connectivity. It doesn't have the power to
create these tables but equally it cannot serve useful traffic without
them and failing the readiness might help to highlight problems in an
environment.